chore: renamed directories to match sklearn, added tests
