+++ shared/lib/linear_model/linearRegression.ts
