+++ shared/lib/linear_model/sgdClassifier.ts
