+++ shared/lib/linear_model/sgdRegressor.ts
