+++ shared/lib/pipeline/pipeline.test.ts
@@ -2,7 +2,7 @@ import { assert } from 'chai'
 import { Pipeline, makePipeline } from './pipeline'
 import { tensor1d } from '@tensorflow/tfjs-core'
 import { tensorEqual } from '../utils'
-import { LinearRegression } from '../estimators/linearRegression'
+import { LinearRegression } from '../linear_model/linearRegression'
 import { SimpleImputer } from '../impute/simpleImputer'
 import { MinMaxScaler } from '../preprocessing/minMaxScaler'
 import { describe, it } from 'mocha'
